ParallelLogger-package {ParallelLogger}	R Documentation

ParallelLogger: Support for Parallel Computation, Logging, and Function Automation

Description

Support for parallel computation with progress bar, and option to stop or proceed on errors. Also provides logging to console and disk, and the logging persists in the parallel threads. Additional functions support function call automation with delayed execution (e.g. for executing functions in parallel).

Author(s)

Maintainer: Martijn Schuemie schuemie@ohdsi.org

Authors:

Marc Suchard

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

https://ohdsi.github.io/ParallelLogger/

https://github.com/OHDSI/ParallelLogger

Report bugs at https://github.com/OHDSI/ParallelLogger/issues

[Package ParallelLogger version 3.1.0 Index]