addDefaultErrorReportLogger {ParallelLogger}	R Documentation

Add the default error report logger

Description

Add the default error report logger

Usage

addDefaultErrorReportLogger(
  fileName = file.path(getwd(), "errorReportR.txt"),
  name = "DEFAULT_ERRORREPORT_LOGGER"
)

Arguments

fileName
The name of the file to write to.

name
A name for the logger.

Details

Creates a logger that writes to a file using the "FATAL" threshold and the layoutErrorReport layout. The file will be overwritten if it is older than 60 seconds. The user will be notified that the error report has been created, and where to find it.

[Package ParallelLogger version 3.1.0 Index]