addDefaultFileLogger {ParallelLogger}	R Documentation

Add the default file logger

Description

Add the default file logger

Usage

addDefaultFileLogger(fileName, name = "DEFAULT_FILE_LOGGER")

Arguments

fileName
The name of the file to write to.

name
A name for the logger.

Details

Creates a logger that writes to a file using the "TRACE" threshold and the layoutParallel layout. The output can be viewed with the built-in log viewer that can be started using launchLogViewer.

[Package ParallelLogger version 3.1.0 Index]