clusterRequire {ParallelLogger}	R Documentation

Require a package in the cluster

Description

Calls the require function in each node of the cluster.

Usage

clusterRequire(cluster, package)

Arguments

cluster
The cluster object.

package
The name of the package to load in all nodes.

[Package ParallelLogger version 3.0.1 Index]