createFileAppender {ParallelLogger}	R Documentation

Create file appender

Description

Create file appender

Usage

createFileAppender(
  layout = layoutParallel,
  fileName,
  overwrite = FALSE,
  expirationTime = 60
)

Arguments

layout
The layout to be used by the appender.

fileName
The name of the file to write to.

overwrite
Overwrite the file if it is older than the expiration time?

expirationTime
Expiration time in seconds

Details

Creates an appender that will write to a file.

[Package ParallelLogger version 3.0.1 Index]