excludeFromList {ParallelLogger}	R Documentation

Exclude variables from a list of objects of the same type

Description

Exclude variables from a list of objects of the same type

Usage

excludeFromList(x, exclude)

Arguments

x
A list of objects of the same type.

exclude
A character vector of names of variables to exclude.

[Package ParallelLogger version 3.1.0 Index]