layoutStackTrace {ParallelLogger}	R Documentation

Logging layout with stack trace

Description

A layout function to be used with an appender. This layout adds the stack trace to the message.

Usage

layoutStackTrace(level, message)

Arguments

level
The level of the message (e.g. "INFO")

message
The message to layout.

[Package ParallelLogger version 3.1.0 Index]