logFatal {ParallelLogger}	R Documentation

Log a message at the FATAL level

Description

Log a message at the FATAL level

Usage

logFatal(...)

Arguments

...
Zero or more objects which can be coerced to character (and which are pasted together with no separator).

Details

Log a message at the specified level. The message will be sent to all the registered loggers. This function is be automatically called when an error occurs, and should not be called directly. Use stop() instead.

[Package ParallelLogger version 3.1.0 Index]