matchInList {ParallelLogger}	R Documentation

In a list of object of the same type, find those that match the input

Description

In a list of object of the same type, find those that match the input

Usage

matchInList(x, toMatch)

Arguments

x
A list of objects of the same type.

toMatch
The object to match.

Details

Typically, toMatch will contain a subset of the variables that are in the objects in the list. Any object matching all variables in toMatch will be included in the result.

Value

A list of objects that match the toMatch object.

Examples

x <- list(
  a = list(name = "John", age = 25, gender = "M"),
  b = list(name = "Mary", age = 24, gender = "F")
)

matchInList(x, list(name = "Mary"))

# $a
# $a$name
# [1] "John"
#
# $a$age
# [1] 25
#
#
# $b
# $b$name
# [1] "Mary"
#
# $b$age
# [1] 24

[Package ParallelLogger version 3.1.0 Index]