unregisterLogger {ParallelLogger}	R Documentation

Unregister a logger

Description

Unregister a logger

Usage

unregisterLogger(x, silent = FALSE)

Arguments

x
Can either be an integer (e.g. 2 to remove the second logger), the name of the logger, or the logger object itself.

silent
If TRUE, no warning will be issued if the logger is not found.

Details

Unregisters a logger from the logging system.

Value

Returns TRUE if the logger was removed.

Examples

appender <- createConsoleAppender(layout = layoutTimestamp)

logger <- createLogger(name = "SIMPLE",
                       threshold = "INFO",
                       appenders = list(appender))
registerLogger(logger)                      
logTrace("This event is below the threshold (INFO)")
logInfo("Hello world")                       
unregisterLogger("SIMPLE")                     

[Package ParallelLogger version 3.1.0 Index]