launchLogViewer {ParallelLogger}	R Documentation

Launch the log viewer Shiny app

Description

Launch the log viewer Shiny app

Usage

launchLogViewer(logFileName)

Arguments

logFileName
Name of the log file to view.

Details

Launches a Shiny app that allows the user to view a log file created using the default file logger. Use addDefaultFileLogger to start the default file logger.

Examples

# Create a log file:
logFile <- file.path(tempdir(), "log.txt")
addDefaultFileLogger(logFile)
logInfo("Hello world")

# Launch the log file viewer (only if in interactive mode):
if (interactive()) {
  launchLogViewer(logFile)
}

# Delete the log file:
unlink(logFile)

[Package ParallelLogger version 3.1.0 Index]